requirements.txt
SoundFile
numpy
torch
torchaudio
torchvision
sounddevice
Cython
librosa
sentencepiece==0.1.91
protobuf==3.12.0
nemo_toolkit[all]==1.0.0b3
deepspeech
jiwer
tabulate
